feat(cashier): display records from API

+++ app/src/main/java/com/dicoding/InsightCart/ui/Cashier/Records/ParentItem.kt
@@ -1,3 +1,3 @@
 package com.dicoding.InsightCart.ui.Cashier.Records
 
-data class ParentItem(val id : String ,val date : String, val total : Int , val mList : List<ChildItem>)
+data class ParentItem(val id : String ,val date : String, val total : Double , val mList : List<ChildItem>)
